描画エリアをコンポーネントに切り出しました。
